requirements.txt
rich
pyyaml
prettyconf
